==> 07d_Regression_between_Veg_and_Nest.Rmd <==
---
title: "07d_Regression_between_Veg_and_Nest"
output: html_document
date: "2023-10-24"
---
#4. Regression Models: 
```{r package loading}
library(betareg)
library(StepBeta)
library(ggstats)
library(usdm)
```
```{r data preparation + collineraity problem check}
#Data Preparation for following model building:
rapid_dat10.model=rapid_dat10
#Convert 0s and 1s of nest presence and nestling survival in order to build beta regression models later.
#the equation is from: https://psycnet.apa.org/doiLanding?doi=10.1037%2F1082-989X.11.1.54
N=length(rapid_dat10$presence)
rapid_dat10.model$presence=(rapid_dat10$presence*(N-1)+0.5)/N
rapid_dat10.model$survival=(rapid_dat10$survival*(N-1)+0.5)/N

#Check collinearity problems of predictor variables
rapid_dat10.vif<-as.data.frame(rapid_dat10.model[,c(19:31)])
vifcor(rapid_dat10.vif,th=0.9)
vifstep(rapid_dat10.vif,th=10)
#No collineraity problems
```
# select the top ranked model based on AIC (any models within 2 delta AIC)
mod_pres<-mod_tab[mod_tab$Response=="Presence",]
form_pres<-mod_pres[mod_pres$AIC==min(mod_pres$AIC),]$Function

mod_surv<-mod_tab[mod_tab$Response=="Success",]
form_surv<-mod_surv[mod_surv$AIC==min(mod_surv$AIC),]$Function

```{r skip}
#test plot(skip)
ggplot(rapid_dat10.model,aes(y=presence,x=patens_pct,color=factor(gerardii_pres)))+
  geom_point()+
  stat_smooth(method="gam",se=FALSE)
rapid_dat10.model.plot<-rapid_dat10.model%>%
  melt(id=c("id","presence","survival"),na.rm = TRUE)%>%
  as.data.frame()

ggplot(rapid_dat10.model.plot,aes(y=presence,x=value))+
        facet_wrap(~variable)+
     geom_point()+
    stat_smooth(method="gam",se=FALSE)

rapid_dat10.high<-rapid_dat10[,c(1,19:33,64:65)]%>%
  melt(id=c("id","presence"),na.rm = TRUE)%>%
  as.data.frame()
rapid_dat10.high.unmelt<-rapid_dat10[,c(1:2,20:22,102)]
rapid_dat10.high.unmelt<-rapid_dat10.high.unmelt[rapid_dat10.high.unmelt$presence>0.9,]
rapid_dat10.high<-rapid_dat10.high[rapid_dat10.high$presence>0.9,]
ggplot(rapid_dat10.high[rapid_dat10.high$year==2022,],aes(y=value,x=variable))+
         geom_violin()+
         geom_line(aes(group=id,color=id))+
        theme(legend.position = "none")
```
```{r betareg survival}

#summary the global model using all predictor variables (percentage)
summary(fullmod_survival_pct<-betareg(survival~alt_tall_pct+distichlis_pct+gerardii_pct+patens_pct+alt_short_pct+phrag_pct+low_marsh_pct+high_marsh_pct+brackish_border_pct+saltmarsh_border_pct+upland_pct+trees_pct+water_pct+Lat+I(Lat^2),data=rapid_dat10.model))

#use stepmod function to simplify the global model
stepmod_survival_pct<-StepBeta(fullmod_survival_pct,k=2)
#check the output (saltmarsh_border_pct is not significant)
summary(stepmod_survival_pct)
#record the simplified model
simplified_surv_pct_Lat2<-betareg(survival ~ alt_tall_pct+ distichlis_pct+gerardii_pct +patens_pct + alt_short_pct+ phrag_pct +low_marsh_pct+ high_marsh_pct + brackish_border_pct  + upland_pct + water_pct  + Lat+I(Lat^2), data = rapid_dat10.model )
#record the simplified model with Lat^2 removed for further comparison
simplified_surv_pct<-betareg(survival ~ alt_tall_pct+ distichlis_pct+gerardii_pct +patens_pct + alt_short_pct+ phrag_pct +low_marsh_pct+ high_marsh_pct + brackish_border_pct + upland_pct+ water_pct  + Lat, data = rapid_dat10.model )
```
```{r betareg presence}
##summary the global model using all predictor variables (percentage)
summary(fullmod_pres_pct<-betareg(presence~alt_tall_pct+distichlis_pct+gerardii_pct+patens_pct+alt_short_pct+phrag_pct+low_marsh_pct+high_marsh_pct+brackish_border_pct+saltmarsh_border_pct+trees_pct+water_pct+upland_pct+Lat+I(Lat^2),data=rapid_dat10.model))
#use stepmod function to simplify the global model
stepmod_pres_pct<-StepBeta(fullmod_pres_pct,k=2)
#check the output (saltmarsh_border_pct is not significant)
stepmod_pres_pct

#record the simplified model
simplified_pres_pct_Lat2<-betareg(formula =  presence ~  alt_tall_pct + gerardii_pct+patens_pct + alt_short_pct +phrag_pct  + high_marsh_pct +trees_pct+Lat +I(Lat^2),data = rapid_dat10.model)
#record the simplified model with Lat^2 removed for further comparison
simplified_pres_pct<-betareg(formula =  presence ~  alt_tall_pct + gerardii_pct+patens_pct + alt_short_pct +phrag_pct  + high_marsh_pct +trees_pct+Lat,data = rapid_dat10.model)

```
```{r }
ggcoef_model(fullmod_survival_pct)
ggcoef_model(simplified_surv_pct)
ggcoef_model(simplified_surv_pct_Lat2)
ggcoef_model(fullmod_pres_pct)
ggcoef_model(simplified_pres_pct)
ggcoef_model(simplified_pres_pct_Lat2)
```
```{r model list presence }
## Mod List
## 1. Select model 
#------------------------------------------
## Start a list of potential models
mod_list_pres1<-list()
# How much variation does each variable explain relative to each other?
#alt_tall
mod_list_pres1[[1]]<-betareg(presence~alt_tall_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_pres1[[2]]<-betareg(presence~alt_tall_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#distichlis
mod_list_pres1[[3]]<-betareg(presence~distichlis_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_pres1[[4]]<-betareg(presence~distichlis_pct+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
#gerardii
mod_list_pres1[[5]]<-betareg(presence~gerardii_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_pres1[[6]]<-betareg(presence~gerardii_pct+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
#patens
mod_list_pres1[[7]]<-betareg(presence~patens_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_pres1[[8]]<-betareg(presence~patens_pct+Lat+I(Lat^2), 
                         data=rapid_dat10.model
)
#alt_short
mod_list_pres1[[9]]<-betareg(presence~alt_short_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_pres1[[10]]<-betareg(presence~alt_short_pct+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
#phrag
mod_list_pres1[[11]]<-betareg(presence~phrag_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_pres1[[12]]<-betareg(presence~phrag_pct+Lat+I(Lat^2),
                          data=rapid_dat10.model
                          )
#HIGHMARSH
mod_list_pres1[[13]]<-betareg(presence~high_marsh_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_pres1[[14]]<-betareg(presence~high_marsh_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#Barckish Border
mod_list_pres1[[15]]<-betareg(presence~brackish_border_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_pres1[[16]]<-betareg(presence~brackish_border_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#Upland

mod_list_pres1[[17]]<-betareg(presence~upland_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_pres1[[18]]<-betareg(presence~upland_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#Trees
mod_list_pres1[[19]]<-betareg(presence~trees_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_pres1[[20]]<-betareg(presence~trees_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#Water
mod_list_pres1[[21]]<-betareg(presence~water_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_pres1[[22]]<-betareg(presence~water_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#LowMarsh
mod_list_pres1[[23]]<-betareg(presence~low_marsh_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_pres1[[24]]<-betareg(presence~low_marsh_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#SaltmarshBorder
mod_list_pres1[[25]]<-betareg(presence~saltmarsh_border_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_pres1[[26]]<-betareg(presence~saltmarsh_border_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )

mod_list_pres1[[27]]<- simplified_pres_pct
mod_list_pres1[[28]]<- fullmod_pres_pct
mod_list_pres1[[29]]<- simplified_pres_pct_Lat2

# store model selection factors in a table
mod_tab_habstr<-data.frame(Response=rep("Presence",length(mod_list_pres1)),
                           Model_Comparison=rep("Veg Species Occurrence",length(mod_list_pres1)),
                           Model_Name=rep(NA,length(mod_list_pres1)),
                           Function=rep(NA,length(mod_list_pres1)),
                           AIC=rep(NA,length(mod_list_pres1)),
                           dAIC=rep(NA,length(mod_list_pres1)))

mod_tab_habstr[1,"Model_Name"]<-"alt tall pres/abs"
mod_tab_habstr[1,"Function"]<-deparse1(mod_list_pres1[[1]]$formula)
mod_tab_habstr[1,"AIC"]<-AIC(mod_list_pres1[[1]])

mod_tab_habstr[2,"Model_Name"]<-"alt tall pct"
mod_tab_habstr[2,"Function"]<-deparse1(mod_list_pres1[[2]]$formula)
mod_tab_habstr[2,"AIC"]<-AIC(mod_list_pres1[[2]])

mod_tab_habstr[3,"Model_Name"]<-"distichlis pres/abs"
mod_tab_habstr[3,"Function"]<-deparse1(mod_list_pres1[[3]]$formula)
mod_tab_habstr[3,"AIC"]<-AIC(mod_list_pres1[[3]])

mod_tab_habstr[4,"Model_Name"]<-"distichlis pct"
mod_tab_habstr[4,"Function"]<-deparse1(mod_list_pres1[[4]]$formula)
mod_tab_habstr[4,"AIC"]<-AIC(mod_list_pres1[[4]])

mod_tab_habstr[5,"Model_Name"]<-"gerardii pres/abs"
mod_tab_habstr[5,"Function"]<-deparse1(mod_list_pres1[[5]]$formula)
mod_tab_habstr[5,"AIC"]<-AIC(mod_list_pres1[[5]])

mod_tab_habstr[6,"Model_Name"]<-"gerardii pct"
mod_tab_habstr[6,"Function"]<-deparse1(mod_list_pres1[[6]]$formula)
mod_tab_habstr[6,"AIC"]<-AIC(mod_list_pres1[[6]])

mod_tab_habstr[7,"Model_Name"]<-"patens pres/abs"
mod_tab_habstr[7,"Function"]<-deparse1(mod_list_pres1[[7]]$formula)
mod_tab_habstr[7,"AIC"]<-AIC(mod_list_pres1[[7]])

mod_tab_habstr[8,"Model_Name"]<-"patens pct"
mod_tab_habstr[8,"Function"]<-deparse1(mod_list_pres1[[8]]$formula)
mod_tab_habstr[8,"AIC"]<-AIC(mod_list_pres1[[8]])

mod_tab_habstr[9,"Model_Name"]<-"alt short pres/abs"
mod_tab_habstr[9,"Function"]<-deparse1(mod_list_pres1[[9]]$formula)
mod_tab_habstr[9,"AIC"]<-AIC(mod_list_pres1[[9]])

mod_tab_habstr[10,"Model_Name"]<-"alt short pct"
mod_tab_habstr[10,"Function"]<-deparse1(mod_list_pres1[[10]]$formula)
mod_tab_habstr[10,"AIC"]<-AIC(mod_list_pres1[[10]])

mod_tab_habstr[11,"Model_Name"]<-"phrag pres/abs"
mod_tab_habstr[11,"Function"]<-deparse1(mod_list_pres1[[11]]$formula)
mod_tab_habstr[11,"AIC"]<-AIC(mod_list_pres1[[11]])

mod_tab_habstr[12,"Model_Name"]<-"phrag pct"
mod_tab_habstr[12,"Function"]<-deparse1(mod_list_pres1[[12]]$formula)
mod_tab_habstr[12,"AIC"]<-AIC(mod_list_pres1[[12]])

mod_tab_habstr[13,"Model_Name"]<-"high marsh pres/abs"
mod_tab_habstr[13,"Function"]<-deparse1(mod_list_pres1[[13]]$formula)
mod_tab_habstr[13,"AIC"]<-AIC(mod_list_pres1[[13]])

mod_tab_habstr[14,"Model_Name"]<-"high marsh pct"
mod_tab_habstr[14,"Function"]<-deparse1(mod_list_pres1[[14]]$formula)
mod_tab_habstr[14,"AIC"]<-AIC(mod_list_pres1[[14]])

mod_tab_habstr[15,"Model_Name"]<-"brackish border pres/abs"
mod_tab_habstr[15,"Function"]<-deparse1(mod_list_pres1[[15]]$formula)
mod_tab_habstr[15,"AIC"]<-AIC(mod_list_pres1[[15]])

mod_tab_habstr[16,"Model_Name"]<-"brackish border pct"
mod_tab_habstr[16,"Function"]<-deparse1(mod_list_pres1[[16]]$formula)
mod_tab_habstr[16,"AIC"]<-AIC(mod_list_pres1[[16]])

mod_tab_habstr[17,"Model_Name"]<-"upland pres/abs"
mod_tab_habstr[17,"Function"]<-deparse1(mod_list_pres1[[17]]$formula)
mod_tab_habstr[17,"AIC"]<-AIC(mod_list_pres1[[17]])

mod_tab_habstr[18,"Model_Name"]<-"upland pct"
mod_tab_habstr[18,"Function"]<-deparse1(mod_list_pres1[[18]]$formula)
mod_tab_habstr[18,"AIC"]<-AIC(mod_list_pres1[[18]])

mod_tab_habstr[19,"Model_Name"]<-"trees pres/abs"
mod_tab_habstr[19,"Function"]<-deparse1(mod_list_pres1[[19]]$formula)
mod_tab_habstr[19,"AIC"]<-AIC(mod_list_pres1[[19]])

mod_tab_habstr[20,"Model_Name"]<-"trees pct"
mod_tab_habstr[20,"Function"]<-deparse1(mod_list_pres1[[20]]$formula)
mod_tab_habstr[20,"AIC"]<-AIC(mod_list_pres1[[20]])

mod_tab_habstr[21,"Model_Name"]<-"water pres/abs"
mod_tab_habstr[21,"Function"]<-deparse1(mod_list_pres1[[21]]$formula)
mod_tab_habstr[21,"AIC"]<-AIC(mod_list_pres1[[21]])

mod_tab_habstr[22,"Model_Name"]<-"water pct"
mod_tab_habstr[22,"Function"]<-deparse1(mod_list_pres1[[22]]$formula)
mod_tab_habstr[22,"AIC"]<-AIC(mod_list_pres1[[22]])

mod_tab_habstr[23,"Model_Name"]<-"low marsh pres/abs"
mod_tab_habstr[23,"Function"]<-deparse1(mod_list_pres1[[23]]$formula)
mod_tab_habstr[23,"AIC"]<-AIC(mod_list_pres1[[23]])

mod_tab_habstr[24,"Model_Name"]<-"low marsh pct"
mod_tab_habstr[24,"Function"]<-deparse1(mod_list_pres1[[24]]$formula)
mod_tab_habstr[24,"AIC"]<-AIC(mod_list_pres1[[24]])

mod_tab_habstr[25,"Model_Name"]<-"saltmarsh border pres/abs"
mod_tab_habstr[25,"Function"]<-deparse1(mod_list_pres1[[25]]$formula)
mod_tab_habstr[25,"AIC"]<-AIC(mod_list_pres1[[25]])

mod_tab_habstr[26,"Model_Name"]<-"saltmarsh border pct"
mod_tab_habstr[26,"Function"]<-deparse1(mod_list_pres1[[26]]$formula)
mod_tab_habstr[26,"AIC"]<-AIC(mod_list_pres1[[26]])

mod_tab_habstr[27,"Model_Name"]<-"simplified model with only Lat"
mod_tab_habstr[27,"Function"]<-deparse1(mod_list_pres1[[27]]$formula)
mod_tab_habstr[27,"AIC"]<-AIC(mod_list_pres1[[27]])

mod_tab_habstr[28,"Model_Name"]<-"global model"
mod_tab_habstr[28,"Function"]<-deparse1(mod_list_pres1[[28]]$formula)
mod_tab_habstr[28,"AIC"]<-AIC(mod_list_pres1[[28]])

mod_tab_habstr[29,"Model_Name"]<-"simplified model with Lat and Lat^2"
mod_tab_habstr[29,"Function"]<-deparse1(mod_list_pres1[[29]]$formula)
mod_tab_habstr[29,"AIC"]<-AIC(mod_list_pres1[[29]])
mod_tab_habstr<-group_by(mod_tab_habstr,Response)%>%mutate(dAIC=AIC-min(AIC))%>%
  ungroup()%>%
  arrange(Response,Model_Comparison,dAIC)


# compare models
mod_tab<-mod_tab_habstr%>%
  group_by(Response)%>%
  mutate(ΔAIC_within_Model_Comparison=round(dAIC,1),
         Overall_ΔAIC_across_Comparisons=round(AIC-min(AIC),1))%>%
  arrange(Response,Model_Comparison,dAIC)%>%
  ungroup()%>%
  dplyr::select(-dAIC) 
mod_tab

  write.csv(mod_tab,paste0(path_out,"Final_outputs/Model_Results/veg_presence_model_selection_table",".csv"), row.names = F)

```
```{r model list survival}
## Mod List
## 1. Select model 
#------------------------------------------
## Start a list of potential models
mod_list_surv1<-list()
# How much variation does each variable explain relative to each other?
#alt_tall
mod_list_surv1[[1]]<-betareg(survival~alt_tall_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_surv1[[2]]<-betareg(survival~alt_tall_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#distichlis
mod_list_surv1[[3]]<-betareg(survival~distichlis_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_surv1[[4]]<-betareg(survival~distichlis_pct+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
#gerardii
mod_list_surv1[[5]]<-betareg(survival~gerardii_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_surv1[[6]]<-betareg(survival~gerardii_pct+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
#patens
mod_list_surv1[[7]]<-betareg(survival~patens_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_surv1[[8]]<-betareg(survival~patens_pct+Lat+I(Lat^2), 
                         data=rapid_dat10.model
)
#alt_short
mod_list_surv1[[9]]<-betareg(survival~alt_short_pres+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
mod_list_surv1[[10]]<-betareg(survival~alt_short_pct+Lat+I(Lat^2), 
                         data=rapid_dat10.model
                         )
#phrag
mod_list_surv1[[11]]<-betareg(survival~phrag_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[12]]<-betareg(survival~phrag_pct+Lat+I(Lat^2),
                          data=rapid_dat10.model
                          )
#HIGHMARSH
mod_list_surv1[[13]]<-betareg(survival~high_marsh_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[14]]<-betareg(survival~high_marsh_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#Barckish Border
mod_list_surv1[[15]]<-betareg(survival~brackish_border_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[16]]<-betareg(survival~brackish_border_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
#Non-saltmarsh

mod_list_surv1[[17]]<-betareg(survival~upland_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[18]]<-betareg(survival~upland_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[19]]<-betareg(survival~water_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[20]]<-betareg(survival~water_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[21]]<-betareg(survival~low_marsh_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[22]]<-betareg(survival~low_marsh_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[23]]<-betareg(survival~saltmarsh_border_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[24]]<-betareg(survival~saltmarsh_border_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )

mod_list_surv1[[25]]<- simplified_surv_pct
mod_list_surv1[[26]]<- fullmod_survival_pct
mod_list_surv1[[27]]<- simplified_surv_pct_Lat2
mod_list_surv1[[28]]<-betareg(survival~trees_pres+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
mod_list_surv1[[29]]<-betareg(survival~trees_pct+Lat+I(Lat^2), 
                          data=rapid_dat10.model
                          )
# store model selection factors in a table
mod_tab_habstr<-data.frame(Response=rep("survival",length(mod_list_surv1)),
                           Model_Comparison=rep("Veg Species Occurrence",length(mod_list_surv1)),
                           Model_Name=rep(NA,length(mod_list_surv1)),
                           Function=rep(NA,length(mod_list_surv1)),
                           AIC=rep(NA,length(mod_list_surv1)),
                           dAIC=rep(NA,length(mod_list_surv1)))

mod_tab_habstr[1,"Model_Name"]<-"alt tall pres/abs"
mod_tab_habstr[1,"Function"]<-deparse1(mod_list_surv1[[1]]$formula)
mod_tab_habstr[1,"AIC"]<-AIC(mod_list_surv1[[1]])

mod_tab_habstr[2,"Model_Name"]<-"alt tall pct"
mod_tab_habstr[2,"Function"]<-deparse1(mod_list_surv1[[2]]$formula)
mod_tab_habstr[2,"AIC"]<-AIC(mod_list_surv1[[2]])

mod_tab_habstr[3,"Model_Name"]<-"distichlis pres/abs"
mod_tab_habstr[3,"Function"]<-deparse1(mod_list_surv1[[3]]$formula)
mod_tab_habstr[3,"AIC"]<-AIC(mod_list_surv1[[3]])

mod_tab_habstr[4,"Model_Name"]<-"distichlis pct"
mod_tab_habstr[4,"Function"]<-deparse1(mod_list_surv1[[4]]$formula)
mod_tab_habstr[4,"AIC"]<-AIC(mod_list_surv1[[4]])

mod_tab_habstr[5,"Model_Name"]<-"gerardii pres/abs"
mod_tab_habstr[5,"Function"]<-deparse1(mod_list_surv1[[5]]$formula)
mod_tab_habstr[5,"AIC"]<-AIC(mod_list_surv1[[5]])

mod_tab_habstr[6,"Model_Name"]<-"gerardii pct"
mod_tab_habstr[6,"Function"]<-deparse1(mod_list_surv1[[6]]$formula)
mod_tab_habstr[6,"AIC"]<-AIC(mod_list_surv1[[6]])

mod_tab_habstr[7,"Model_Name"]<-"patens pres/abs"
mod_tab_habstr[7,"Function"]<-deparse1(mod_list_surv1[[7]]$formula)
mod_tab_habstr[7,"AIC"]<-AIC(mod_list_surv1[[7]])

mod_tab_habstr[8,"Model_Name"]<-"patens pct"
mod_tab_habstr[8,"Function"]<-deparse1(mod_list_surv1[[8]]$formula)
mod_tab_habstr[8,"AIC"]<-AIC(mod_list_surv1[[8]])

mod_tab_habstr[9,"Model_Name"]<-"alt short pres/abs"
mod_tab_habstr[9,"Function"]<-deparse1(mod_list_surv1[[9]]$formula)
mod_tab_habstr[9,"AIC"]<-AIC(mod_list_surv1[[9]])

mod_tab_habstr[10,"Model_Name"]<-"alt short pct"
mod_tab_habstr[10,"Function"]<-deparse1(mod_list_surv1[[10]]$formula)
mod_tab_habstr[10,"AIC"]<-AIC(mod_list_surv1[[10]])

mod_tab_habstr[11,"Model_Name"]<-"phrag pres/abs"
mod_tab_habstr[11,"Function"]<-deparse1(mod_list_surv1[[11]]$formula)
mod_tab_habstr[11,"AIC"]<-AIC(mod_list_surv1[[11]])

mod_tab_habstr[12,"Model_Name"]<-"phrag pct"
mod_tab_habstr[12,"Function"]<-deparse1(mod_list_surv1[[12]]$formula)
mod_tab_habstr[12,"AIC"]<-AIC(mod_list_surv1[[12]])

mod_tab_habstr[13,"Model_Name"]<-"high marsh pres/abs"
mod_tab_habstr[13,"Function"]<-deparse1(mod_list_surv1[[13]]$formula)
mod_tab_habstr[13,"AIC"]<-AIC(mod_list_surv1[[13]])

mod_tab_habstr[14,"Model_Name"]<-"high marsh pct"
mod_tab_habstr[14,"Function"]<-deparse1(mod_list_surv1[[14]]$formula)
mod_tab_habstr[14,"AIC"]<-AIC(mod_list_surv1[[14]])

mod_tab_habstr[15,"Model_Name"]<-"brackish border pres/abs"
mod_tab_habstr[15,"Function"]<-deparse1(mod_list_surv1[[15]]$formula)
mod_tab_habstr[15,"AIC"]<-AIC(mod_list_surv1[[15]])

mod_tab_habstr[16,"Model_Name"]<-"brackish border pct"
mod_tab_habstr[16,"Function"]<-deparse1(mod_list_surv1[[16]]$formula)
mod_tab_habstr[16,"AIC"]<-AIC(mod_list_surv1[[16]])

mod_tab_habstr[17,"Model_Name"]<-"upland pres/abs"
mod_tab_habstr[17,"Function"]<-deparse1(mod_list_surv1[[17]]$formula)
mod_tab_habstr[17,"AIC"]<-AIC(mod_list_surv1[[17]])

mod_tab_habstr[18,"Model_Name"]<-"upland pct"
mod_tab_habstr[18,"Function"]<-deparse1(mod_list_surv1[[18]]$formula)
mod_tab_habstr[18,"AIC"]<-AIC(mod_list_surv1[[18]])

mod_tab_habstr[19,"Model_Name"]<-"water pres/abs"
mod_tab_habstr[19,"Function"]<-deparse1(mod_list_surv1[[19]]$formula)
mod_tab_habstr[19,"AIC"]<-AIC(mod_list_surv1[[19]])

mod_tab_habstr[20,"Model_Name"]<-"water pct"
mod_tab_habstr[20,"Function"]<-deparse1(mod_list_surv1[[20]]$formula)
mod_tab_habstr[20,"AIC"]<-AIC(mod_list_surv1[[20]])

mod_tab_habstr[21,"Model_Name"]<-"low marsh pres/abs"
mod_tab_habstr[21,"Function"]<-deparse1(mod_list_surv1[[21]]$formula)
mod_tab_habstr[21,"AIC"]<-AIC(mod_list_surv1[[21]])

mod_tab_habstr[22,"Model_Name"]<-"low marsh pct"
mod_tab_habstr[22,"Function"]<-deparse1(mod_list_surv1[[22]]$formula)
mod_tab_habstr[22,"AIC"]<-AIC(mod_list_surv1[[22]])

mod_tab_habstr[23,"Model_Name"]<-"saltmarsh border pres/abs"
mod_tab_habstr[23,"Function"]<-deparse1(mod_list_surv1[[22]]$formula)
mod_tab_habstr[23,"AIC"]<-AIC(mod_list_surv1[[23]])

mod_tab_habstr[24,"Model_Name"]<-"saltmarsh border pct"
mod_tab_habstr[24,"Function"]<-deparse1(mod_list_surv1[[24]]$formula)
mod_tab_habstr[24,"AIC"]<-AIC(mod_list_surv1[[24]])

mod_tab_habstr[25,"Model_Name"]<-"simplified model with only Lat"
mod_tab_habstr[25,"Function"]<-deparse1(mod_list_surv1[[25]]$formula)
mod_tab_habstr[25,"AIC"]<-AIC(mod_list_surv1[[25]])

mod_tab_habstr[26,"Model_Name"]<-"global model"
mod_tab_habstr[26,"Function"]<-deparse1(mod_list_surv1[[26]]$formula)
mod_tab_habstr[26,"AIC"]<-AIC(mod_list_surv1[[26]])

mod_tab_habstr[27,"Model_Name"]<-"simplified model with Lat and Lat^2"
mod_tab_habstr[27,"Function"]<-deparse1(mod_list_surv1[[27]]$formula)
mod_tab_habstr[27,"AIC"]<-AIC(mod_list_surv1[[27]])

mod_tab_habstr[28,"Model_Name"]<-"trees pres/abs"
mod_tab_habstr[28,"Function"]<-deparse1(mod_list_surv1[[28]]$formula)
mod_tab_habstr[28,"AIC"]<-AIC(mod_list_surv1[[28]])

mod_tab_habstr[29,"Model_Name"]<-"trees pct"
mod_tab_habstr[29,"Function"]<-deparse1(mod_list_surv1[[29]]$formula)
mod_tab_habstr[29,"AIC"]<-AIC(mod_list_surv1[[29]])
mod_tab_habstr<-group_by(mod_tab_habstr,Response)%>%mutate(dAIC=AIC-min(AIC))%>%
  ungroup()%>%
  arrange(Response,Model_Comparison,dAIC)


# compare models
mod_tab<-mod_tab_habstr%>%
  group_by(Response)%>%
  mutate(ΔAIC_within_Model_Comparison=round(dAIC,1),
         Overall_ΔAIC_across_Comparisons=round(AIC-min(AIC),1))%>%
  arrange(Response,Model_Comparison,dAIC)%>%
  ungroup()%>%
  dplyr::select(-dAIC) 
mod_tab

  write.csv(mod_tab,paste0(path_out,"Final_outputs/Model_Results/veg_survival_model_selection_table",".csv"), row.names = F)

```
AIC(b<-betareg(presence~alt_short_pres*distichlis_pres+
              patens_pct*gerardii_pct+
              phrag_pres*patens_pct+
              alt_short_pres*alt_tall_pct+
              phrag_pres*trees_pct+`non-saltmarsh_pct`+
              high_marsh_pct+brackish_border_pct+
              Lat,data=rapid_dat10.model))
AIC(betareg(presence~alt_short_pres+distichlis_pres+
              patens_pct+gerardii_pct+
              phrag_pres+alt_tall_pct+
              trees_pct+`non-saltmarsh_pct`+
              high_marsh_pct+brackish_border_pct+
              Lat,data=rapid_dat10.model))
if(!file.exists(paste0(path_out,"Final_outputs/Model_Results/veg_model_selection_table_",ab_type,".csv"))){
  write.csv(mod_tab%>%dplyr::select(-Function),paste0(path_out,"Final_outputs/Model_Results/veg_model_selection_table_",ab_type,".csv"), row.names = F)
}
```
